backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
requests
email-validator
passlib[bcrypt]
bcrypt==4.0.1
pytest
httpx